+++ packages/band/src/quantiles/Quantiles.tsx
@@ -3,13 +3,13 @@ import {
     OpacityMotion,
     useDisabledKeys,
     useScales,
-    getIdKeySets,
+    useIdsKeys,
     isBandAxisScale,
     TooltipDataComponent,
     useProcessedData,
 } from '@chsk/core'
 import { useQuantilePreparedData } from './context'
-import { ReactNode, createElement, useMemo } from 'react'
+import { createElement, useMemo } from 'react'
 import { BoxAndWhiskers } from './BoxAndWhiskers'
 import { QuantilePreparedDataItem, QuantilesProps } from './types'
 import { isQuantileProcessedData } from './predicates'
@@ -34,12 +34,8 @@ export const Quantiles = ({
     const data = preparedData.data
     const { disabledKeys, firstRender } = useDisabledKeys()
     const horizontal = isBandAxisScale(scales.y)
+    const { idSet, keySet } = useIdsKeys(ids, keys, preparedData)
     if (!isQuantileProcessedData(processedData)) return null
-
-    const { idSet, keySet } = useMemo(
-        () => getIdKeySets(ids, keys, preparedData),
-        [ids, keys, preparedData]
-    )
 
     const allKeys = preparedData.keys
     const boxStyles = useMemo(
@@ -55,7 +51,7 @@ export const Quantiles = ({
         [allKeys, middleStyle, colorScale]
     )
 
-    const result: Array<ReactNode> = preparedData.keys
+    const result = preparedData.keys
         .map((k, i) => {
             if (!keySet.has(k)) return null
             const visible = !disabledKeys.has(k)