+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### Changed
 
+-   [core] replaced `getIdKeySets` with hook `useIdsKeys`; reduces the need for
+    `useMemo` in add-on packages
 -   [band] props for `BarsLabels` to match `DendrogramLeafLabels`
 -   [all] added conditional exports to `package.json` files, specifying entry points for 'import' and 'require'
 -   [all] updated dependencies in `package.json` files to enable recent